+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Added
 
 - use translation function to allow localization across plugin UI labels and messages
+- new GPT-5.6 models in the settings list
 
 ## 2.28.5 - 2026-08-20